Use the helpers in the proxy (and drop InterpreterLocatorHelper).
