+++ src/client/pythonEnvironments/discovery/locators/helpers.ts
@@ -1,19 +1,10 @@
 import * as fsapi from 'fs-extra';
-import { inject } from 'inversify';
 import * as path from 'path';
 import { Uri } from 'vscode';
 import { traceError } from '../../../common/logger';
 import { IS_WINDOWS } from '../../../common/platform/constants';
 import { IFileSystem } from '../../../common/platform/types';
-import { IPipEnvServiceHelper } from '../../../interpreter/locators/types';
-import {
-    areSameInterpreter,
-    InterpreterType,
-    normalizeInterpreter,
-    PythonInterpreter,
-    updateInterpreter
-} from '../../info';
-import { areSameVersion } from '../../info/pythonVersion';
+import { InterpreterType, PythonInterpreter } from '../../info';
 
 const CheckPythonInterpreterRegEx = IS_WINDOWS ? /^python(\d+(.\d+)?)?\.exe$/ : /^python(\d+(.\d+)?)?$/;
 
@@ -33,36 +24,6 @@ export async function lookForInterpretersInDirectory(pathToCheck: string, _: IFi
     }
 }
 
-export class InterpreterLocatorHelper {
-    constructor(
-        @inject(IFileSystem) private readonly fs: IFileSystem,
-        @inject(IPipEnvServiceHelper) private readonly pipEnvServiceHelper: IPipEnvServiceHelper
-    ) {}
-
-    public async mergeInterpreters(interpreters: PythonInterpreter[]): Promise<PythonInterpreter[]> {
-        const deps = {
-            updateInterpreter,
-            areSameInterpreter: (i1: PythonInterpreter, i2: PythonInterpreter) =>
-                areSameInterpreter(i1, i2, {
-                    areSameVersion,
-                    inSameDirectory: (p1?: string, p2?: string) =>
-                        inSameDirectory(p1, p2, {
-                            arePathsSame: this.fs.arePathsSame,
-                            getPathDirname: path.dirname
-                        })
-                }),
-            normalizeInterpreter: (i: PythonInterpreter) => normalizeInterpreter(i, { normalizePath: path.normalize }),
-            getPipEnvInfo: this.pipEnvServiceHelper.getPipEnvInfo.bind(this.pipEnvServiceHelper)
-        };
-        const merged = mergeInterpreters(interpreters, deps);
-        await Promise.all(
-            // At this point they are independent so we can update them separately.
-            merged.map(async (interp) => updateEnvInfo(interp, deps))
-        );
-        return merged;
-    }
-}
-
 /**
  * Combine env info for matching environments.
  *
@@ -73,7 +34,7 @@ export class InterpreterLocatorHelper {
  * @prop deps.areSameInterpreter - determine if 2 infos match the same env
  * @prop deps.normalizeInterpreter - standardize the given env info
  */
-function mergeInterpreters(
+export function mergeInterpreters(
     interpreters: PythonInterpreter[],
     deps: {
         areSameInterpreter(i1: PythonInterpreter, i2: PythonInterpreter): boolean;
@@ -103,7 +64,7 @@ function mergeInterpreters(
  * @prop deps.arePathsSame - determine if two filenames point to the same file
  * @prop deps.getPathDirname - (like `path.dirname`)
  */
-function inSameDirectory(
+export function inSameDirectory(
     path1: string | undefined,
     path2: string | undefined,
     deps: {
@@ -126,7 +87,7 @@ function inSameDirectory(
  * @param deps - functional dependencies
  * @prop deps.getPipEnvInfo - provides extra pip-specific env info, if applicable
  */
-async function updateEnvInfo(
+export async function updateEnvInfo(
     interp: PythonInterpreter,
     deps: {
         getPipEnvInfo(p: string): Promise<{ workspaceFolder: Uri; envName: string } | undefined>;
